# src/components/User/model.ts
import * as bcrypt from 'bcrypt';
import * as connections from '../../config/connection/connection';
import * as crypto from 'crypto';
import { Document, Schema } from 'mongoose';
import { NextFunction } from 'express';

/**
* @export
* @interface IUserModel
* @extends {Document}
*/
export interface IUserModel extends Document {
  mobile: string;
  email: string;
  password: string;
  passwordResetToken: string;
  passwordResetExpires: Date;

  tokens: AuthToken[];

  profile: {
    isChild: boolean,
    name: string,
    gender: string,
    location: string,
    picture: string
  };
  comparePassword: (password: string) => Promise < boolean > ;
  gravatar: (size: number) => string;
}

export type AuthToken = {
  accessToken: string,
  kind: string
};

/**
* @swagger
* components:
*  schemas:
*    UserSchema:
*      properties:
*        id:
*          type: string
*        mobile:
*          type: string
*        email:
*          type: string
*        password:
*          type: string
*        passwordResetToken:
*          type: string
*        passwordResetExpires:
*          type: string
*          format: date
*        tokens:
*          type: array
*        profile:
*          properties:
*            isChild:
*              type: boolean
*            name:
*              type: string
*            gender:
*              type: string
*            location:
*              type: string
*            picture:
*              type: string
*          required:
*            - isChild
*            - name
*      required:
*        - email
*        - password
*      examples:
*        mobile: "0033648347451"
*        email: "1605285140"
*        password: "bob"
*        profile: {
*          "isChild": false,
*          "name": "Julie"
*        }
*    Users:
*      type: array
*      items:
*        $ref: '#/components/schemas/UserSchema'
*/
const UserSchema: Schema = new Schema({
  mobile: {
    type: String,
    trim: true,
    default: ''
  },
  email: {
    type: String,
    unique: true,
    trim: true
  },
  password: String,
  passwordResetToken: String,
  passwordResetExpires: Date,
  tokens: Array,
  profile: {
    isChild: Boolean,
    name: String,
    gender: {
      type: String,
      default: ''
    },
    location: {
      type: String,
      default: ''
    },
    picture: {
      type: String,
      default: ''
    }
  }
}, {
  collection: 'usermodel',
  versionKey: false
}).pre('save', async function (next: NextFunction): Promise < void > {
  const user: any = this; // tslint:disable-line

  if (!user.isModified('password')) {
    return next();
  }

  try {
    const salt: string = await bcrypt.genSalt(10);

    const hash: string = await bcrypt.hash(user.password, salt);

    user.password = hash;
    next();
  } catch (error) {
    return next(error);
  }
});

/**
* Method for comparing passwords
*/
UserSchema.methods.comparePassword = async function (candidatePassword: string): Promise < boolean > {
  try {
    const match: boolean = await bcrypt.compare(candidatePassword, this.password);

    return match;
  } catch (error) {
    return error;
  }
};

/**
* Helper method for getting user's gravatar.
*/
UserSchema.methods.gravatar = function (size: number): string {
  if (!size) {
    size = 200;
  }
  if (!this.email) {
    return `https://gravatar.com/avatar/?s=${size}&d=retro`;
  }
  const md5: string = crypto.createHash('md5').update(this.email).digest('hex');

  return `https://gravatar.com/avatar/${md5}?s=${size}&d=retro`;
};

export default connections.db.model < IUserModel > ('UserModel', UserSchema);